Move model_config to common (#48)
Addresses triton-inference-server/server#2759:
 many of the functions in model_config.h are generically useful even on the client
side and only depend on the generated protobuf headers. Therefore, it is useful
 to have them more broadly accessible via the common repo.

The header and cc files are copied here from the core repo (except for a few 
functions with non-trivial dependencies that stay in core), with the
 namespace changed to triton::common. A new library 
triton-common-model-config is built. Because many of the functions depend on
the generated protobuf headers, this library is only built when the protobuf part of
 this repo is enabled.

This PR goes with triton-inference-server/core#60.